WebSockets/tests/submissions/Opera/constructor/011.html

<!doctype html>
<title>WebSockets: protocol mismatch</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(3000, 'got no close event');
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/handshake_protocol', 'FOOBAR');
  var gotOpen = false;
  var gotError = false;
  ws.onopen = function(e) {
    debug(e);
    gotOpen = true;
  }
  ws.onerror = function(e) {
    debug(e);
    gotError = true;
  }
  ws.onclose = function(e) {
    debug(e);
    assert_equals(gotOpen, true, 'got open');
    assert_equals(gotError, true, 'got error');
    ws.onclose = assertUnreached;
    setTestTimeout(null);
    setTimeout(end, 50);
  }
});
</script>
